[scripts/fast-reboot] Shutdown remaining containers through systemd
The current implementation has two issues:

1. In case containers from "docker ps" output are ordered in a way that
database is first in the list, the "systemctl stop database" followed by
"docker kill database" will stop all other containers through systemd
and ruin this optimization
2. After "docker kill database" there are lots of errors from daemons
like hostcfgd, system-healthd, caclmgrd, etc. Also it causes those
daemons to hang when received SIGTERM making a delay on following
"systemctl stop database".

In the new implementation, services are implicitelly stopped by systemd
in the order that is correct. If a certain container needs an
optimization that will kill the container instead of stopping it the
container may implement this optimization in its /usr/local/bin/*.sh
script.

It is also more optimal since independent services might be stopped in
parallel.

NOTE: This fix is relevant for regular SONiC image and not for
Kubernetes enabled SONiC image. Kubernetes integration in SONiC might
have this issue still, which might be a design flawn.

Signed-off-by: Stepan Blyschak <stepanb@nvidia.com>